Playing around with Render and Frame buffers, as well as basic Kernels.

+++ src/Camera.cpp
@@ -15,7 +15,7 @@ void Camera::updateMatrix(float FOVdeg, float nearPlane, float farPlane)
 	// Forms a view matrix from a Position(Eye), a Direction(Center) and an Orientation(Which way is Up)
   glm::mat4 view = glm::lookAt(Position, Position + Orientation, Up);
 	// Forms the perspective matrix with a FOV, a canvas size, nearPlan and farPlane 
-  glm::mat4 projection = glm::perspective(glm::radians(FOVdeg), (float)(width/height), nearPlane, farPlane);
+  glm::mat4 projection = glm::perspective(glm::radians(FOVdeg), (static_cast<float>(width) / static_cast<float>(height)), nearPlane, farPlane);
 
 	// Sets new camera matrix
 	cameraMatrix = projection * view;
